+++ semcore/base-components/src/components/ellipsis/index.ts
@@ -1,0 +1,12 @@
+import { useEllipsis, isTextOverflowing } from './useEllipsis';
+import type { EllipsisSettings } from './useEllipsis';
+import { useResizeObserver } from './useResizeObserver';
+
+export {
+  useEllipsis,
+  isTextOverflowing,
+  useResizeObserver,
+};
+export type {
+  EllipsisSettings,
+};