+++ semcore/base-components/src/components/hint/index.ts
@@ -1,0 +1,7 @@
+import { Hint } from './Hint';
+import type { HintPopperProps } from './Hint';
+
+export { Hint };
+export type {
+  HintPopperProps,
+};