+++ semcore/tooltip/src/index.js
@@ -1,2 +1,1 @@
 export { default, Hint, DescriptionTooltip } from './Tooltip';
-export { HintPopper } from './HintPopper';