+++ stories/components/link/docs/examples/links_with_ellipsis.tsx
@@ -6,18 +6,17 @@ import { Text } from '@semcore/typography';
 import React from 'react';
 
 const Demo = () => {
+  const lintRef = React.useRef(null);
   return (
     <Flex>
       <Text flex='0 0 auto'>Sep 3</Text>
       <Divider mx={4} orientation='vertical' />
-      <Link w={500} wMin={0} href='#'>
+      <Link w={500} wMin={0} href='#' ref={lintRef}>
         <Flex alignItems='center'>
-          <Link.Text w='100%' inline noWrap>
-            <Text w='100%' inline noWrap>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit. Atque iusto, sed!
-              Asperiores, consectetur deserunt et ipsam omnis quae repellendus velit veniam.
-              Asperiores dicta dolor ducimus enim fugit laborum minima reprehenderit?
-            </Text>
+          <Link.Text w={480} tag={Text} display='inline-block' ellipsis={true} hintProps={{ triggerRef: lintRef }}>
+            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Atque iusto, sed!
+            Asperiores, consectetur deserunt et ipsam omnis quae repellendus velit veniam.
+            Asperiores dicta dolor ducimus enim fugit laborum minima reprehenderit?
           </Link.Text>
           <Link.Addon>
             <LinkExternalM />