+++ semcore/base-components/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.2.0] - 2025-08-08
+
+### Added
+
+- `useEllipsis` hook.
+- `Hint` component.
+
 ## [16.1.2] - 2025-07-23
 
 ### Changed